+++ src/utils/clean/__init__.py
@@ -1,1 +1,5 @@
 """Functions for cleaning data in ways that may make assumptions"""
+
+from utils.clean._core import clean_data
+
+__all__ = ["clean_data"]
