+++ src/utils/standardize/__init__.py
@@ -4,3 +4,7 @@ The DataSource class is subclassed for each unique source of data. For
 more information, see TODO and CONTRIBUTING.md for how to add additional
 states
 """
+
+from utils.standardize._core import standardize_state
+
+__all__ = ["standardize_state"]
